+++ src/converse-controlbox.js
@@ -466,7 +466,7 @@
                     if (!this.validate()) { return; }
 
                     const form_data = new FormData(ev.target);
-                    _converse.session.save({
+                    _converse.config.save({
                         'trusted': form_data.get('trusted') && true || false,
                         'storage': form_data.get('trusted') ? 'local' : 'session'
                     });
@@ -585,7 +585,7 @@
             });
 
             _converse.on('clearSession', () => {
-                if (_converse.session.get('trusted')) {
+                if (_converse.config.get('trusted')) {
                     const chatboxes = _.get(_converse, 'chatboxes', null);
                     if (!_.isNil(chatboxes)) {
                         const controlbox = chatboxes.get('controlbox');